+++ pages/observation.py
@@ -10,6 +10,7 @@ from dash import callback_context
 import numpy as np
 from data.real_life_meaning_mapping import real_life_meaning_mapping
 import plotly.graph_objects as go
+from src.trendplot import create_aggregated_time_series_plot
 
 # from functions.visualization import map_fig
 
@@ -30,48 +31,6 @@ df["price_adjusted"] = df["price"].str.extract(r'([0-9.]+)', expand = False).ast
 df["bathroom_adjusted"] = df["bathrooms_text"].str.extract(r'([0-9.]+)', expand = False).astype(float)
 
 simulated = pd.read_csv('data/raw/simulated.csv')
-
-# Function to create a time-series plot
-def create_aggregated_time_series_plot(df, y_variable=None):
-    if y_variable is None:
-        y_variable = 'Daily Price'
-    y_name = y_variable
-    y_variable = real_life_meaning_mapping[y_name]['column_name']
-    df = df.copy()
-    year_quarter = df['quarter'].str.split('-', expand=True)
-    
-    # Convert year and quarter into a Period object
-    df['quarter'] = pd.PeriodIndex(year=year_quarter[0].astype(int), 
-                                   quarter=year_quarter[1].astype(int), 
-                                   freq='Q')
-    df['quarter'] = df['quarter'].dt.strftime('%Y-Q%q')
-    # Aggregating the data
-    aggregated_df = df.groupby('quarter')[y_variable].mean().reset_index()
-    median_aggregated_df = df.groupby('quarter')[y_variable].median().reset_index()
-
-    # Creating an empty figure and adding both mean and median as separate traces
-    fig = go.Figure()
-
-    # Add Mean trace
-    fig.add_trace(go.Scatter(x=aggregated_df['quarter'], y=aggregated_df[y_variable],
-                             mode='lines', name='Mean'))
-
-    # Adding median trend line
-    fig.add_trace(go.Scatter(x=median_aggregated_df['quarter'], y=median_aggregated_df[y_variable],
-                             mode='lines', name='Median'))
-    # Center the title
-    fig.update_layout(
-        title={
-            'text': f'Trend of the Metric {y_name}',
-            'y':0.9,
-            'x':0.5,
-            'xanchor': 'center',
-            'yanchor': 'top'
-        }
-    )
-    fig.update_xaxes(title_text='Quarter', tickangle=-45)
-    fig.update_yaxes(title_text=f'Average {y_name}')
-    return fig
 
 SIDEBAR_STYLE = {
     "top": 42,
